Add a new DAG that tests the status of node pool are reported correctly
This change adds a new DAG which automates the process of going through
the lifecycle of a node pool and verifies whether the node pool status
is reported correctly.

This DAG requires an existing cluster. It will create a node-pool under
the specified cluster, and clean it up after the tests.

---------

Co-authored-by: Alfred Yu <alfred.yu@cienet.com>